requirements.txt
aiogram==2.25.1
aiohttp==3.9.1
aiomysql==0.2.0
python-dotenv==1.0.0
